+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Don't persist master secret key in `KeyManager` and derive account level keys on initialization by @TheCharlatan ([#322](https://github.com/farcaster-project/farcaster-core/pull/322))
 - Rename `DoubleKeys` into `SwapRoleKeys` by @h4sh3d ([#324](https://github.com/farcaster-project/farcaster-core/pull/324))
 
+### Removed
+
+- Stand alone reveal proof message is removed and part of Alice and Bob reveal parameters by @h4sh3d ([#325](https://github.com/farcaster-project/farcaster-core/pull/325))
+
 ## [0.6.3] - 2022-12-28
 
 ### Added